Make Header for Landing Page in TOP
Created Header section that contains logo description and images.
images is not fitting to my expectation - extra space is Created
for entire container of the image while trying to fit the image to
the size of the container. Otherwise the image only takes half the space
of the container. Everything is quite up to expectation but not perfect